More work on Python FFI - loading from Arrow datastreams
