+++ hnsw_redux/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-async-trait = "0.1.83"
 bytemuck = "1.20.0"
 datafusion = { workspace = true }
 enum_dispatch = "0.3.13"
@@ -17,4 +16,5 @@ rand = "0.8.5"
 rayon = "1.10.0"
 serde = { version = "1.0.210", features = ["derive"] }
 serde_json = "1.0.128"
+thiserror = "2.0.11"
 unroll = "0.1.5"
